+++ offchain-modules/packages/app-cli/src/index.ts
@@ -12,7 +12,7 @@ import { sigCmd } from './sigServer';
 
 export const program = commander.program;
 
-const version = '0.0.34';
+const version = '0.0.38';
 
 async function main() {
   program