+++ offchain-modules/packages/x/src/db/stat.ts
@@ -1,5 +1,6 @@
 import { Connection, MoreThan } from 'typeorm';
 import { CkbBurn } from './entity/CkbBurn';
+import { CollectorEthUnlock } from './entity/EthUnlock';
 
 export class StatDb {
   constructor(private connection: Connection) {}
@@ -9,6 +10,15 @@ export class StatDb {
     return await this.connection.getRepository(CkbBurn).find({
       where: {
         createdAt: MoreThan(new Date(Date.now() - interval * 1000)),
+      },
+    });
+  }
+
+  // get CollectorEthUnlocks status manual-review
+  async getManualReviewCollectorEthUnlocks(): Promise<CollectorEthUnlock[]> {
+    return await this.connection.getRepository(CollectorEthUnlock).find({
+      where: {
+        status: 'manual-review',
       },
     });
   }