+++ offchain-modules/packages/x/src/handlers/ckb.ts
@@ -3,6 +3,7 @@ import { serializeMultisigScript } from '@ckb-lumos/common-scripts/lib/secp256k1
 import { key } from '@ckb-lumos/hd';
 import { encodeToAddress, sealTransaction, TransactionSkeletonType } from '@ckb-lumos/helpers';
 import { RPC } from '@ckb-lumos/rpc';
+import { BigNumber } from 'bignumber.js';
 import { Reader } from 'ckb-js-toolkit';
 import * as lodash from 'lodash';
 import { BtcAsset, ChainType, EosAsset, EthAsset, getAsset, TronAsset } from '../ckb/model/asset';
@@ -16,7 +17,7 @@ import { forceBridgeRole } from '../config';
 import { ForceBridgeCore } from '../core';
 import { CkbDb, KVDb } from '../db';
 import { CollectorCkbMint } from '../db/entity/CkbMint';
-import { ICkbBurn, ICkbMint, MintedRecord, MintedRecords } from '../db/model';
+import { ICkbBurn, ICkbMint, IEthUnlock, MintedRecord, MintedRecords } from '../db/model';
 import { asserts, nonNullable } from '../errors';
 import { BridgeMetricSingleton, txTokenInfo } from '../metric/bridge-metric';
 import { createAsset, MultiSigMgr } from '../multisig/multisig-mgr';
@@ -29,6 +30,7 @@ import {
   uint8ArrayToString,
 } from '../utils';
 import { logger } from '../utils/logger';
+import { getCachedAssetAVGPrice } from '../utils/price';
 import { getAssetTypeByAsset } from '../xchain/tron/utils';
 import Transaction = CKBComponents.Transaction;
 import TransactionWithStatus = CKBComponents.TransactionWithStatus;
@@ -106,6 +108,28 @@ export class CkbHandler {
       if (BigInt(burn.amount) <= BigInt(burn.bridgeFee))
         throw new Error('Unexpected error: burn amount less than bridge fee');
       const unlockAmount = (BigInt(burn.amount) - BigInt(burn.bridgeFee)).toString();
+      const collectorEthUnlock: IEthUnlock = {
+        ckbTxHash: burn.ckbTxHash,
+        asset: burn.asset,
+        amount: unlockAmount,
+        recipientAddress: burn.recipientAddress,
+      };
+      const individualAuditThreshold =
+        ForceBridgeCore.config.audit && ForceBridgeCore.config.audit!.individualAuditThreshold
+          ? ForceBridgeCore.config.audit!.individualAuditThreshold
+          : '100000';
+      const asset = ForceBridgeCore.config.eth.assetWhiteList.find((asset) => asset.address === burn.asset);
+      if (!asset) throw new Error('asset not in white list');
+      const price = await getCachedAssetAVGPrice(asset.symbol);
+      if (
+        new BigNumber(unlockAmount)
+          .multipliedBy(price)
+          .div(Math.pow(10, asset.decimal))
+          .gt(new BigNumber(individualAuditThreshold))
+      ) {
+        collectorEthUnlock.status = 'manual-review';
+        collectorEthUnlock.message = `Amount of ${unlockAmount} ${asset.symbol} is greater than the threshold of ${individualAuditThreshold}`;
+      }
       switch (burn.chain) {
         case ChainType.BTC:
           await this.db.createBtcUnlock([
@@ -119,14 +143,7 @@ export class CkbHandler {
           ]);
           break;
         case ChainType.ETH:
-          await this.db.createCollectorEthUnlock([
-            {
-              ckbTxHash: burn.ckbTxHash,
-              asset: burn.asset,
-              amount: unlockAmount,
-              recipientAddress: burn.recipientAddress,
-            },
-          ]);
+          await this.db.createCollectorEthUnlock([collectorEthUnlock]);
           break;
         case ChainType.EOS:
           await this.db.createEosUnlock([